Add metrics
