+++ src/transports/metrics.ts
@@ -3,26 +3,7 @@ import { TransportGenerics } from '.'
 import { EndpointContext } from '../adapter'
 import { calculateCacheKey, calculateFeedId } from '../cache'
 import { AdapterConfig } from '../config'
-import { requestDurationBuckets } from '../metrics/constants'
 import { InputParameters } from '../validation'
-
-// Data Provider Metrics
-export const dataProviderMetricsLabel = (providerStatusCode?: number, method = 'get') => ({
-  provider_status_code: providerStatusCode,
-  method: method.toUpperCase(),
-})
-
-export const dataProviderRequests = new client.Counter({
-  name: 'data_provider_requests',
-  help: 'The number of http requests that are made to a data provider',
-  labelNames: ['method', 'provider_status_code'] as const,
-})
-
-export const dataProviderRequestDurationSeconds = new client.Histogram({
-  name: 'data_provider_request_duration_seconds',
-  help: 'A histogram bucket of the distribution of data provider request durations',
-  buckets: requestDurationBuckets,
-})
 
 // Websocket Metrics
 export const connectionErrorLabels = (message: string) => ({
